Extract validation code
