+++ src/main.js
@@ -1,9 +1,6 @@
 const { parse } = require("./ebnf-parser");
-const {
-  createDocumentation,
-  documentStyle,
-  validateEbnf
-} = require("./report-builder");
+const { createDocumentation, documentStyle } = require("./report-builder");
+const { validateEbnf } = require("./validate");
 const { version } = require("../package.json");
 const {
   getReferences,
